Refactor priest settings, move buffing to base class
