+++ AIO/Helpers/Enums.cs
@@ -54,6 +54,7 @@ namespace WholesomeTBCAIO.Helpers
             { "Shadow", Specs.PriestShadow },
             { "Party Shadow", Specs.PriestShadowParty },
             { "Party Holy", Specs.PriestHolyParty },
+            { "Raid Holy", Specs.PriestRaidParty },
         };
         private readonly static Dictionary<string, Specs> RogueSpecs = new Dictionary<string, Specs>()
         {
@@ -111,6 +112,7 @@ namespace WholesomeTBCAIO.Helpers
             PriestShadow,
             PriestShadowParty,
             PriestHolyParty,
+            PriestRaidParty,
             // Rogue
             RogueCombat,
             RogueCombatParty,
