+++ AIO/Rotations/Priest/Priest.cs
@@ -121,6 +121,30 @@ namespace WholesomeTBCAIO.Rotations.Priest
                     .ToList();
                 if (needRes.Count > 0 && cast.OnFocusUnit(Resurrection, needRes[0]))
                     return;
+
+                List<AIOPartyMember> aliveMembers = AIOParty.Group
+                    .FindAll(m => m.IsAlive && m.GetDistance < 60)
+                    .ToList();
+
+                // Party Cure Disease
+                WoWPlayer needCureDisease = aliveMembers
+                    .Find(m => ToolBox.HasDiseaseDebuff(m.Name));
+                if (needCureDisease != null && cast.OnFocusUnit(CureDisease, needCureDisease))
+                    return;
+
+                // Party Dispel Magic
+                WoWPlayer needDispelMagic = aliveMembers
+                    .Find(m => ToolBox.HasMagicDebuff(m.Name));
+                if (needDispelMagic != null && cast.OnFocusUnit(DispelMagic, needDispelMagic))
+                    return;
+
+                // OOC Inner Fire
+                if (settings.UseInnerFire
+                    && cast.BuffSelf(InnerFire))
+                    return;
+
+                if (BuffParty())
+                    return;
             }
         }
 
@@ -146,7 +170,7 @@ namespace WholesomeTBCAIO.Rotations.Priest
                 && cast.Buff(aliveMembers, PrayerOfShadowProtection, 17029))
                 return true;
 
-            // Prayer Of Shadow Protection
+            // Shadow Protection
             if (settings.PartyShadowProtection
                 && !settings.PartyPrayerOfShadowProtection
                 && cast.Buff(aliveMembers, ShadowProtection))
